checksum field for public_key table
That's neccassary since it's not possible to add a index to a 'text' typed field. At least MySQL dies with an error when running the migration.
To provide uniqueness of the content, the easiest way is to add a checksum field with fixed length, containing the checksum of the key content and a unique index on that field.